requirements.txt
tensorflow>=2.16.0
protobuf>=3.20.3,<7
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
Pillow>=9.0.0
streamlit>=1.28.0
ultralytics>=8.0.0
PyYAML>=6.0
pandas>=1.5.0
h5py>=3.0.0
